basic working prototype

+++ script.js
@@ -1,16 +1,38 @@
-const MONTHS = ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"];
-
-const getTodayDate = () => {
-    let date = new Date();
-    let day = date.getDate();
-    let month = date.getMonth() + 1;
-    let monthName = MONTHS[month - 1];
-    let year = date.getFullYear();
-
-    let dateInWords = `${monthName} ${day}, ${year}`;
-    let dateInNum = `${year}-${month}-${day}`;
-    return [dateInWords, dateInNum];
+clearAllNotesBtn.onclick = () => {
+    let userAction = confirm("Do you really want to clear all your notes?");
+    if (userAction) {
+        localStorage.clear();
+    };
+    createTimeNotes();
 }
 
-localStorage.clear();
+newNoteBtn.onclick = () => {
+    document.body.scrollTop = document.documentElement.scrollTop = 0;
 
+    if (newNoteBtn.innerText == 'New Note') {
+        newNoteBtn.innerText = 'Show All Notes';
+        createNewNoteForm();
+        document.getElementById('formNoteHeading').focus()
+        clearAllNotesBtn.disabled = true;
+    }
+    else {
+        newNoteBtn.innerText = 'New Note';
+        notesSection.textContent = '';
+        createTimeNotes()
+        clearAllNotesBtn.disabled = false;
+    }
+};
+
+let note = {
+    title: "Hello adfasdfasdfthis is a message",
+    text: "klfdjaldfja sdfas dflkasj dflasj fjaskdfl jasldfk aslkdf las jflasj fd"
+}
+
+
+addNoteInLocalStorage(note);
+addNoteInLocalStorage(note);
+addNoteInLocalStorage(note);
+
+
+
+createTimeNotes();
